+++ .cspell.json
@@ -4,8 +4,26 @@
   "ignorePaths": ["**/*.json", "**/*.css", "node_modules", "**/*.log"],
   "useGitignore": true,
   "language": "en",
-  "words": ["dataurl", "devpool", "fkey", "mswjs", "outdir", "servedir", "supabase", "typebox", "ubiquity-os",
-		"smee", "tomlify", "hono", "cfworker", "pavlovcik", "ghijklmnopqrstuvwxyz", "GHIJKLMNOPQRSTUVWXYZ", "ubiquityos"],
+  "words": [
+    "dataurl",
+    "devpool",
+    "fkey",
+    "mswjs",
+    "outdir",
+    "servedir",
+    "supabase",
+    "typebox",
+    "ubiquity-os",
+    "smee",
+    "tomlify",
+    "hono",
+    "cfworker",
+    "workerd",
+    "pavlovcik",
+    "ghijklmnopqrstuvwxyz",
+    "GHIJKLMNOPQRSTUVWXYZ",
+    "ubiquityos"
+  ],
   "dictionaries": ["typescript", "node", "software-terms"],
   "import": ["@cspell/dict-typescript/cspell-ext.json", "@cspell/dict-node/cspell-ext.json", "@cspell/dict-software-terms"],
   "ignoreRegExpList": ["[0-9a-fA-F]{6}"]
