+++ src/adapters/cloudflare-worker.ts
@@ -1,0 +1,3 @@
+import { app } from "../kernel";
+
+export default app;
