+++ tests/dispatch.test.ts
@@ -142,7 +142,16 @@ describe("handleEvent", () => {
     const payloadString = JSON.stringify(payload);
     const signature = calculateSignature(payloadString, secret);
 
-    const req = new Request("http://localhost:8080", {
+    process.env = {
+      ENVIRONMENT: "production",
+      APP_WEBHOOK_SECRET: secret,
+      APP_ID: "1",
+      APP_PRIVATE_KEY: "1234",
+      OPENAI_API_KEY: "token",
+    };
+
+    const app = (await import("../src/kernel")).app;
+    const res = await app.request("http://localhost:8080", {
       method: "POST",
       headers: {
         "x-github-event": "issue_comment.created",
@@ -153,39 +162,8 @@ describe("handleEvent", () => {
       body: payloadString,
     });
 
-    const worker = (await import("../src/worker")).default;
-    // I didn't find a better option to make sure that the non-awaited ctx.waitUntil is resolved
-    // eslint-disable-next-line no-async-promise-executor
-    const waitUntilPromise = new Promise<void>(async (resolve) => {
-      const res = await worker.fetch(
-        req,
-        {
-          ENVIRONMENT: "production",
-          APP_WEBHOOK_SECRET: secret,
-          APP_ID: "1",
-          APP_PRIVATE_KEY: "1234",
-          PLUGIN_CHAIN_STATE: {} as KVNamespace,
-          OPENAI_API_KEY: "token",
-        },
-        {
-          waitUntil<T>(promise: Promise<T>) {
-            promise
-              .then(() => {
-                resolve();
-              })
-              .catch(() => {
-                resolve();
-              });
-          },
-          passThroughOnException() {},
-        }
-      );
-
-      expect(res).toBeTruthy();
-    });
-
-    await waitUntilPromise;
-
+    expect(res).toBeTruthy();
+    // 2 calls means the execution didn't break
     expect(dispatchWorker).toHaveBeenCalledTimes(2);
 
     dispatchWorker.mockReset();
