+++ tests/main.test.ts
@@ -5,7 +5,7 @@ import { http, HttpResponse } from "msw";
 import { GitHubContext } from "../src/github/github-context";
 import { GitHubEventHandler } from "../src/github/github-event-handler";
 import { getConfig } from "../src/github/utils/config";
-import worker from "../src/worker"; // has to be imported after the mocks
+import { app } from "../src/kernel"; // has to be imported after the mocks
 import { server } from "./__mocks__/node";
 import "./__mocks__/webhooks";
 
@@ -59,24 +59,19 @@ describe("Worker tests", () => {
     );
   });
   it("Should fail on missing env variables", async () => {
-    const req = new Request("http://localhost:8080");
     const consoleSpy = jest.spyOn(console, "error").mockImplementation(() => jest.fn());
-    const res = await worker.fetch(
-      req,
-      {
-        ENVIRONMENT: "production",
-        APP_WEBHOOK_SECRET: "",
-        APP_ID: "",
-        APP_PRIVATE_KEY: "",
-        PLUGIN_CHAIN_STATE: {} as KVNamespace,
-        OPENAI_API_KEY: "token",
-      },
-      {
-        waitUntil: function () {},
-        passThroughOnException: function () {},
-      }
-    );
+    process.env = {
+      ENVIRONMENT: "production",
+      APP_WEBHOOK_SECRET: "",
+      APP_ID: "",
+      APP_PRIVATE_KEY: "",
+      OPENAI_API_KEY: "token",
+    };
+    const res = await app.request("http://localhost:8080", {
+      method: "POST",
+    });
     expect(res.status).toEqual(500);
+    expect(await res.json()).toEqual({ error: "Error: Invalid environment variables" });
     consoleSpy.mockReset();
   });
