+++ wrangler.toml
@@ -1,19 +1,19 @@
 name = "ubiquity-os-kernel"
-main = "src/worker.ts"
+main = "src/adapters/cloudflare-worker.ts"
 compatibility_date = "2023-12-06"
-compatibility_flags = [ "nodejs_compat" ]
+compatibility_flags = ["nodejs_compat"]
 
 # Prefer this syntax due to a bug in Wrangler: https://github.com/cloudflare/workers-sdk/issues/5634
 [env]
- [env.dev]
-  [[env.dev.kv_namespaces]]
-		binding = "PLUGIN_CHAIN_STATE"
-		id = "4f7aadc56bef41a7ae2cc8c0582320b3"
+[env.dev]
+[[env.dev.kv_namespaces]]
+binding = "PLUGIN_CHAIN_STATE"
+id = "4f7aadc56bef41a7ae2cc8c0582320b3"
 
- [env.production]
-  [[env.production.kv_namespaces]]
-		binding = "PLUGIN_CHAIN_STATE"
-		id = "TO_BE_DEFINED"
+[env.production]
+[[env.production.kv_namespaces]]
+binding = "PLUGIN_CHAIN_STATE"
+id = "TO_BE_DEFINED"
 
 # Enables Cloudflare Worker Logs
 [observability]
